+++ src/Controller/ImagineController.php
@@ -59,7 +59,8 @@ class ImagineController
     public function filterAction(Request $request, string $path, string $filter): RedirectResponse
     {
         $path = PathHelper::urlPathToFilePath($path);
-        $resolver = $request->get('resolver');
+        // TODO once we limit `symfony/http-foundation` to 6.4 or newer, use `$request->query->getString()`
+        $resolver = $request->query->has('resolver') ? (string) $request->query->get('resolver') : null;
 
         return $this->createRedirectResponse(function () use ($path, $filter, $resolver, $request) {
             return $this->filterService->getUrlOfFilteredImage(
@@ -84,7 +85,7 @@ class ImagineController
      */
     public function filterRuntimeAction(Request $request, string $hash, string $path, string $filter): RedirectResponse
     {
-        $resolver = $request->get('resolver');
+        $resolver = $request->query->has('resolver') ? (string) $request->query->get('resolver') : null;
         $path = PathHelper::urlPathToFilePath($path);
         $runtimeConfig = $request->query->all('filters');